requirements.txt
langchain
ipykernel
python-dotenv
langchain-community
pypdf
langchain_huggingface
faiss-cpu
bs4
langchain-text-splitters
streamlit
sentence_transformers
langchain_huggingface
langchain_groq
langchain
validators==0.28.1
youtube_transcript_api
unstructured
arxiv
wikipedia
duckduckgo-search
pymupdf
pymysql
crewai
crewai_tools
instructor
docstring-parser
arxiv